Project Codes/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.5.0
python-dotenv==1.0.1
google-generativeai==0.8.3
Werkzeug==3.1.0
PyPDF2==3.0.1
python-docx==1.1.0